+++ lib/flutter_datetime_picker_bdaya.dart
@@ -13,21 +13,21 @@ typedef DateChangedCallback(DateTime time);
 typedef DateCancelledCallback();
 typedef String? StringAtIndexCallBack(int index);
 
-class DatePicker {
+class DatePickerBdaya {
   ///
   /// Display date picker bottom sheet.
   ///
   static Future<DateTime?> showDatePicker(
     BuildContext context, {
-    bool showTitleActions: true,
+    bool showTitleActions = true,
     DateTime? minTime,
     DateTime? maxTime,
     DateChangedCallback? onChanged,
     DateChangedCallback? onConfirm,
     DateCancelledCallback? onCancel,
-    locale: LocaleType.en,
+    locale = LocaleType.en,
     DateTime? currentTime,
-    DatePickerThemeBDaya? theme,
+    DatePickerThemeBdaya? theme,
   }) async {
     return await Navigator.push(
       context,
@@ -55,14 +55,14 @@ class DatePicker {
   ///
   static Future<DateTime?> showTimePicker(
     BuildContext context, {
-    bool showTitleActions: true,
-    bool showSecondsColumn: true,
+    bool showTitleActions = true,
+    bool showSecondsColumn = true,
     DateChangedCallback? onChanged,
     DateChangedCallback? onConfirm,
     DateCancelledCallback? onCancel,
-    locale: LocaleType.en,
+    locale = LocaleType.en,
     DateTime? currentTime,
-    DatePickerThemeBDaya? theme,
+    DatePickerThemeBdaya? theme,
   }) async {
     return await Navigator.push(
       context,
@@ -89,13 +89,13 @@ class DatePicker {
   ///
   static Future<DateTime?> showTime12hPicker(
     BuildContext context, {
-    bool showTitleActions: true,
+    bool showTitleActions = true,
     DateChangedCallback? onChanged,
     DateChangedCallback? onConfirm,
     DateCancelledCallback? onCancel,
-    locale: LocaleType.en,
+    locale = LocaleType.en,
     DateTime? currentTime,
-    DatePickerThemeBDaya? theme,
+    DatePickerThemeBdaya? theme,
   }) async {
     return await Navigator.push(
       context,
@@ -121,15 +121,15 @@ class DatePicker {
   ///
   static Future<DateTime?> showDateTimePicker(
     BuildContext context, {
-    bool showTitleActions: true,
+    bool showTitleActions = true,
     DateTime? minTime,
     DateTime? maxTime,
     DateChangedCallback? onChanged,
     DateChangedCallback? onConfirm,
     DateCancelledCallback? onCancel,
-    locale: LocaleType.en,
+    locale = LocaleType.en,
     DateTime? currentTime,
-    DatePickerThemeBDaya? theme,
+    DatePickerThemeBdaya? theme,
   }) async {
     return await Navigator.push(
       context,
@@ -157,13 +157,13 @@ class DatePicker {
   ///
   static Future<DateTime?> showPicker(
     BuildContext context, {
-    bool showTitleActions: true,
+    bool showTitleActions = true,
     DateChangedCallback? onChanged,
     DateChangedCallback? onConfirm,
     DateCancelledCallback? onCancel,
-    locale: LocaleType.en,
+    locale = LocaleType.en,
     BasePickerModel? pickerModel,
-    DatePickerThemeBDaya? theme,
+    DatePickerThemeBdaya? theme,
   }) async {
     return await Navigator.push(
       context,
@@ -188,13 +188,13 @@ class _DatePickerRoute<T> extends PopupRoute<T> {
     this.onChanged,
     this.onConfirm,
     this.onCancel,
-    DatePickerThemeBDaya? theme,
+    DatePickerThemeBdaya? theme,
     this.barrierLabel,
     this.locale,
     RouteSettings? settings,
     BasePickerModel? pickerModel,
   })  : this.pickerModel = pickerModel ?? DatePickerModel(),
-        this.theme = theme ?? DatePickerThemeBDaya(),
+        this.theme = theme ?? DatePickerThemeBdaya(),
         super(settings: settings);
 
   final bool? showTitleActions;
@@ -202,7 +202,7 @@ class _DatePickerRoute<T> extends PopupRoute<T> {
   final DateChangedCallback? onConfirm;
   final DateCancelledCallback? onCancel;
   final LocaleType? locale;
-  final DatePickerThemeBDaya theme;
+  final DatePickerThemeBdaya theme;
   final BasePickerModel pickerModel;
 
   @override
@@ -290,7 +290,7 @@ class _DatePickerState extends State<_DatePickerComponent> {
 
   @override
   Widget build(BuildContext context) {
-    DatePickerThemeBDaya theme = widget.route.theme;
+    DatePickerThemeBdaya theme = widget.route.theme;
     return GestureDetector(
       child: AnimatedBuilder(
         animation: widget.route.animation!,
@@ -323,7 +323,7 @@ class _DatePickerState extends State<_DatePickerComponent> {
     }
   }
 
-  Widget _renderPickerView(DatePickerThemeBDaya theme) {
+  Widget _renderPickerView(DatePickerThemeBdaya theme) {
     Widget itemView = _renderItemView(theme);
     if (widget.route.showTitleActions == true) {
       return Column(
@@ -338,7 +338,7 @@ class _DatePickerState extends State<_DatePickerComponent> {
 
   Widget _renderColumnView(
     ValueKey key,
-    DatePickerThemeBDaya theme,
+    DatePickerThemeBdaya theme,
     StringAtIndexCallBack stringAtIndexCB,
     ScrollController scrollController,
     int layoutProportion,
@@ -393,7 +393,7 @@ class _DatePickerState extends State<_DatePickerComponent> {
     );
   }
 
-  Widget _renderItemView(DatePickerThemeBDaya theme) {
+  Widget _renderItemView(DatePickerThemeBdaya theme) {
     return Container(
       color: theme.backgroundColor,
       child: Directionality(
@@ -468,7 +468,7 @@ class _DatePickerState extends State<_DatePickerComponent> {
   }
 
   // Title View
-  Widget _renderTitleActionsView(DatePickerThemeBDaya theme) {
+  Widget _renderTitleActionsView(DatePickerThemeBdaya theme) {
     final done = _localeDone();
     final cancel = _localeCancel();
 
@@ -532,15 +532,13 @@ class _BottomPickerLayout extends SingleChildLayoutDelegate {
   _BottomPickerLayout(
     this.progress,
     this.theme, {
-    this.itemCount,
     this.showTitleActions,
     this.bottomPadding = 0,
   });
 
   final double progress;
-  final int? itemCount;
   final bool? showTitleActions;
-  final DatePickerThemeBDaya theme;
+  final DatePickerThemeBdaya theme;
   final double bottomPadding;
 
   @override